+++ CHANGELOG.md
@@ -17,15 +17,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added Features and Improvements 🙌:
 - `Similarity` class derives now from `sklearn.base.BaseEstimator`
-- `fit_transform` method was added to `Similarity` class
+- `Clustering` class derives now from `sklearn.base.BaseEstimator` and `sklearn.base.BaseMixin`
+- `fit_transform` and `transform`  method were added to `Similarity` class
+- `fit_predict` method was added to `Clustering` class
+- `score` method (for silhouette score) was added to `Clustering` class
 - Added LGTM reports
 - Allow to select the desired precision in the CI directly
 - Improved clusters sorting algorithm to prefer larger clusters
+- `mode='kmedoids'` was added to CI
 
 ### Bugfix 🐛:
 - Fix bug of `load_clusters` in case of a single cluster
 - Fix bug where `mode='kmedoids'` can not be executed with `n_clusters`
 - Fix bug for `mode='kmedoids'`, preventing masking diagonal
+- Prevent interpolation in plot
 
 ### Other changes:
 - Improved test functions of utils module